fix: sh and sleep applications
- fix sh parsing empty command error
- change sleep argument type from u8 to usize

+++ user/sleep.zig
@@ -10,7 +10,7 @@ export fn main(argc: usize, argv: [*:null]const ?[*:0]const u8) noreturn {
         sys.exit(1);
     }
 
-    _ = sys.sleep(std.fmt.parseInt(u8, std.mem.sliceTo(argv[1].?, 0), 0) catch |err| {
+    _ = sys.sleep(std.fmt.parseInt(usize, std.mem.sliceTo(argv[1].?, 0), 0) catch |err| {
         fprint(2, "{s}\n", .{@errorName(err)});
         sys.exit(1);
     });
